+++ lib/ruby_lokalise_api/resources/team.rb
@@ -3,8 +3,8 @@
 module RubyLokaliseApi
   module Resources
     class Team < Base
-      MAIN_PARAMS = [].freeze
-      no_support_for %i[update destroy reload_data]
+      MAIN_PARAMS = %i[team_id].freeze
+      no_support_for %i[update destroy]
     end
   end
 end